Refactored the code to be more modular and to suit my idiosyncratic easthetic whims

+++ deep_q_rl/updates.py
@@ -34,27 +34,6 @@ import numpy as np
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-# The MIT License (MIT)
-
-# Copyright (c) 2014 Sander Dieleman
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
 
 def deepmind_rmsprop(loss_or_grads, params, learning_rate, 
                      rho, epsilon):
@@ -107,20 +86,16 @@ def deepmind_rmsprop(loss_or_grads, params, learning_rate,
     for param, grad in zip(params, grads):
         value = param.get_value(borrow=True)
 
-        acc_grad = theano.shared(np.zeros(value.shape, dtype=value.dtype),
-                             broadcastable=param.broadcastable)
-        acc_grad_new = rho * acc_grad + (1 - rho) * grad
+        acc_grad = theano.shared( np.zeros( value.shape, dtype = value.dtype ),
+                                  broadcastable = param.broadcastable )
+        acc_grad_new = rho * acc_grad + ( 1 - rho ) * grad
 
-        acc_rms = theano.shared(np.zeros(value.shape, dtype=value.dtype),
-                             broadcastable=param.broadcastable)
-        acc_rms_new = rho * acc_rms + (1 - rho) * grad ** 2
+        acc_rms = theano.shared( np.zeros( value.shape, dtype = value.dtype ),
+                                  broadcastable = param.broadcastable )
+        acc_rms_new = rho * acc_rms + ( 1 - rho ) * grad ** 2
 
-
-        updates[acc_grad] = acc_grad_new
-        updates[acc_rms] = acc_rms_new
-
-        updates[param] = (param - learning_rate * 
-                          (grad / 
-                           T.sqrt(acc_rms_new - acc_grad_new **2 + epsilon)))
+        updates[ acc_grad ] = acc_grad_new
+        updates[ acc_rms ] = acc_rms_new
+        updates[ param ] = ( param - learning_rate *  ( grad / T.sqrt( acc_rms_new - acc_grad_new ** 2 + epsilon ) ) )
 
     return updates
